Task Scheduler Solution
We need to first count the frequencies of each task and form a Max Heap. Make a queue that will store the remaining frequency of the task and when it can be again performed. When time becomes equal to the time task can be done we push it again into the heap and again pop the element with the highest frequencies.