Combination Sum Solution
Perform the depth-first search and on every recursion either add the same element again or add the next element until the sum of all elements is greater than the target or all elements are visited in nums. Append those arrays which have the sum equal to the target and return to stop further recursion from there.