Product of Array Except Self Solution
First approach for Time Complexity O(n) and space complexity also O(n) is that in first pass we multiply ith element with previous product and keep on adding in prefix array. In second pass we store postfix. Then we just have to multiply postfix of ith element with its prefix.
But this can be applied without using two extra arrays. We can store prefix first in ans array and then by using a variable storing postfix we can pass in reverse direction and multiply it with respective element in ans array.