Car Fleet Solution
As no car can overtake the one in front so by iterating in reverse order of position whenever a car comes which has the time of reaching the target more than that of the car fleet in front of it, it will form another car fleet. So we have to search for these types of cars and increase the count of the car fleets.